+++ pages/index.tsx
@@ -2,56 +2,52 @@ import Head from "next/head";
 import { useEffect, useState } from "react";
 import { GetStaticPropsResult } from "next";
 
-import AboutMe from "../components/MainPage/AboutMe";
-import Contact from "../components/MainPage/Contact";
-import HeaderComponent from "../components/MainPage/HeaderComponent/index";
-import Portfolio from "../components/MainPage/Portfolio";
-import SkillCard from "../components/MainPage/SkillCard";
-import SideMenu from "../components/MainPage/SideMenu";
-import TopMenu from "../components/MainPage/TopMenu";
-
-import { Container } from "../styles/index/styles";
+import { Container } from "../styles/indexV2/styles";
 import { fetchApi } from "../controllers/utils/fetchDatabase";
+import TopMenu from "../components/MainPage_V2/Menus/TopMenu";
+import SideMenu from "../components/MainPage_V2/Menus/SideMenu";
+import Header from "../components/MainPage_V2/Header/Header";
+import About from "../components/MainPage_V2/About/About";
 
 type HomeProps = {
   content: ContentType;
   portfolio: PortfolioItemType[];
 };
 
-const Home: React.FC<HomeProps> = ({ content, portfolio }) => {
+const HomeV2: React.FC<HomeProps> = ({ content, portfolio }) => {
   const [showSideMenu, setShowSideMenu] = useState<boolean>(false);
   const [isMobile, setIsMobile] = useState<boolean>(false);
 
-  useEffect(() => {
-    const skillCards = document.querySelectorAll(".Card");
-    const skillCardsObserver = new IntersectionObserver(
-      (elements, skillCardsObserver) => {
-        elements.forEach((element) => {
-          if (element.isIntersecting) {
-            if (element.target.classList.contains("front")) {
-              element.target.classList.remove("front");
-            }
-            if (element.target.classList.contains("back")) {
-              element.target.classList.remove("back");
-            }
+  //   useEffect(() => {
+  //     const skillCards = document.querySelectorAll(".Card");
+  //     const skillCardsObserver = new IntersectionObserver(
+  //       (elements, skillCardsObserver) => {
+  //         elements.forEach((element) => {
+  //           if (element.isIntersecting) {
+  //             if (element.target.classList.contains("front")) {
+  //               element.target.classList.remove("front");
+  //             }
+  //             if (element.target.classList.contains("back")) {
+  //               element.target.classList.remove("back");
+  //             }
 
-            skillCardsObserver.unobserve(element.target);
-          }
-        });
-      },
-      { threshold: 0.5 }
-    );
+  //             skillCardsObserver.unobserve(element.target);
+  //           }
+  //         });
+  //       },
+  //       { threshold: 0.5 }
+  //     );
 
-    skillCards.forEach((skillCard) => {
-      skillCardsObserver.observe(skillCard);
-    });
-  }, []);
+  //     skillCards.forEach((skillCard) => {
+  //       skillCardsObserver.observe(skillCard);
+  //     });
+  //   }, []);
 
   useEffect(() => {
     if (isMobile) {
       setShowSideMenu(true);
     } else {
-      const root = document.getElementById("headerFrame");
+      const target = document.getElementById("Header");
 
       const observer = new IntersectionObserver(
         (elements) => {
@@ -63,10 +59,10 @@ const Home: React.FC<HomeProps> = ({ content, portfolio }) => {
             }
           });
         },
-        { rootMargin: "-350px 0px 0px 0px" }
+        { threshold: 0.95 }
       );
 
-      observer.observe(root);
+      observer.observe(target);
       return () => observer.disconnect();
     }
   }, [isMobile]);
@@ -88,7 +84,7 @@ const Home: React.FC<HomeProps> = ({ content, portfolio }) => {
   }, []);
 
   return (
-    <div id="TopPage">
+    <div id="Home">
       <Head>
         <title>Hi! I am Denis Lima</title>
         <link rel="icon" href="/img/favico.ico" />
@@ -103,47 +99,44 @@ const Home: React.FC<HomeProps> = ({ content, portfolio }) => {
         />
       </Head>
       <Container>
-        {!showSideMenu && <TopMenu />}
-        <HeaderComponent />
-        <AboutMe id="AboutMe" content={content?.about} />
-        <SkillCard content={content} />
-        <Portfolio id="Portfolio" portfolio={portfolio} />
-        <Contact id="Contact" email={content?.email} />
-        {showSideMenu && <SideMenu />}
+        <TopMenu className={(isMobile || showSideMenu) && "hide"} />
+        <SideMenu className={showSideMenu && "show"} />
+        <Header />
+        <About />
       </Container>
     </div>
   );
 };
 
-export default Home;
+export default HomeV2;
 
-export async function getStaticProps(
-  _context
-): Promise<GetStaticPropsResult<HomeProps>> {
-  let content = {} as ContentType;
-  let portfolio = [] as PortfolioItemType[];
-  try {
-    const resultContent = await fetchApi(
-      "pageContent",
-      process.env.NEXT_PUBLIC_CONTENT_ID
-    );
-    const resultPortfolio = await fetchApi(
-      "portfolioItems",
-      process.env.NEXT_PUBLIC_PORTFOLIO_ID
-    );
-    if (resultContent) {
-      content = resultContent.data() as ContentType;
-    }
-    if (resultPortfolio) {
-      const portfolioResponse =
-        resultPortfolio.data() as PortfolioFetchResponseType;
-      portfolio = portfolioResponse?.items;
-    }
-  } catch (e) {
-    console.log(e);
-  }
-  return {
-    props: { content, portfolio },
-    revalidate: 24 * 60 * 60, // Every 24h
-  };
-}
+// export async function getStaticProps(
+//   _context
+// ): Promise<GetStaticPropsResult<HomeProps>> {
+//   let content = {} as ContentType;
+//   let portfolio = [] as PortfolioItemType[];
+//   try {
+//     const resultContent = await fetchApi(
+//       "pageContent",
+//       process.env.NEXT_PUBLIC_CONTENT_ID
+//     );
+//     const resultPortfolio = await fetchApi(
+//       "portfolioItems",
+//       process.env.NEXT_PUBLIC_PORTFOLIO_ID
+//     );
+//     if (resultContent) {
+//       content = resultContent.data() as ContentType;
+//     }
+//     if (resultPortfolio) {
+//       const portfolioResponse =
+//         resultPortfolio.data() as PortfolioFetchResponseType;
+//       portfolio = portfolioResponse?.items;
+//     }
+//   } catch (e) {
+//     console.log(e);
+//   }
+//   return {
+//     props: { content, portfolio },
+//     revalidate: 24 * 60 * 60, // Every 24h
+//   };
+// }
